switch metrics to float64, add support of timer and histogram

+++ internal/metrics/collect.go
@@ -58,7 +58,7 @@ func GoRuntimeStats(prefix []string) {
 
 	prefix = append(prefix, "runtime")
 
-	UpdateGauge(append(prefix, "num_goroutines"), float32(runtime.NumGoroutine()))
+	UpdateGauge(append(prefix, "num_goroutines"), int64(runtime.NumGoroutine()))
 
 	var s runtime.MemStats
 	runtime.ReadMemStats(&s)
@@ -73,7 +73,7 @@ func GoRuntimeStats(prefix []string) {
 	// occur simultaneously, and as a result HeapAlloc tends to
 	// change smoothly (in contrast with the sawtooth that is
 	// typical of stop-the-world garbage collectors).
-	UpdateGauge(append(prefix, "heap_alloc"), float32(s.Alloc))
+	UpdateGauge(append(prefix, "heap_alloc"), int64(s.Alloc))
 
 	// Sys is the total bytes of memory obtained from the OS.
 	// Sys is the sum of the XSys fields below. Sys measures the
@@ -82,28 +82,28 @@ func GoRuntimeStats(prefix []string) {
 	// likely that not all of the virtual address space is backed
 	// by physical memory at any given moment, though in general
 	// it all was at some point.
-	UpdateGauge(append(prefix, "sys"), float32(s.Sys))
+	UpdateGauge(append(prefix, "sys"), int64(s.Sys))
 
 	// PauseTotalNs is the cumulative nanoseconds in GC
 	// stop-the-world pauses since the program started.
 	//
 	// During a stop-the-world pause, all goroutines are paused
 	// and only the garbage collector can run.
-	UpdateGauge(append(prefix, "pause_total_ns"), float32(s.PauseTotalNs))
+	UpdateGauge(append(prefix, "pause_total_ns"), int64(s.PauseTotalNs))
 
 	// NumGC is the number of completed GC cycles.
-	UpdateGauge(append(prefix, "num_gc"), float32(s.NumGC))
+	UpdateGauge(append(prefix, "num_gc"), int64(s.NumGC))
 
 	// HeapReleased is bytes of physical memory returned to the OS.
 	//
 	// This counts heap memory from idle spans that was returned
 	// to the OS and has not yet been reacquired for the heap.
-	UpdateGauge(append(prefix, "heap_released"), float32(s.HeapReleased))
+	UpdateGauge(append(prefix, "heap_released"), int64(s.HeapReleased))
 
 	// HeapObjects is the number of allocated heap objects.
 	//
 	// Like HeapAlloc, this increases as objects are allocated and
 	// decreases as the heap is swept and unreachable objects are
 	// freed.
-	UpdateGauge(append(prefix, "heap_objects"), float32(s.HeapReleased))
+	UpdateGauge(append(prefix, "heap_objects"), int64(s.HeapReleased))
 }
